requirements.txt
langgraph>=0.1.17
langchain>=0.2.16
langchain-openai>=0.1.25
langchain-core>=0.2.41
python-dotenv>=1.0.1
pydantic>=2.9.0
